back_utils/report-generation/requirements.txt
markdown2==2.5.3
openai==1.93.0
python-dotenv==1.1.1
weasyprint==65.1
markdown2==2.5.3
